Revert tweetUrl to unit test
